Cast to int to silence warnings

+++ src/core/implementation/signal_postponer.cpp
@@ -34,9 +34,9 @@ void SignalPostponer::setPatiece(const std::chrono::milliseconds& p)
 void SignalPostponer::notify()
 {
     if (m_lazinessTimer.isActive() == false)        // lazy timer will be launched for the first time?
-        m_patienceTimer.start(m_patience.count());  // start patience timer also
+        m_patienceTimer.start(static_cast<int>(m_patience.count()));  // start patience timer also
 
-    m_lazinessTimer.start(m_delay.count());
+    m_lazinessTimer.start(static_cast<int>(m_delay.count()));
 }
